+++ baybe/constraints/base.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
+from collections.abc import Sequence
 from typing import TYPE_CHECKING, Any, ClassVar
 
 import pandas as pd
@@ -17,7 +18,6 @@ from baybe.serialization import (
     get_base_structure_hook,
     unstructure_base,
 )
-from baybe.utils.numerical import DTypeFloatTorch
 
 if TYPE_CHECKING:
     from torch import Tensor
@@ -148,7 +148,7 @@ class ContinuousConstraint(Constraint, ABC):
         return [1.0] * len(self.parameters)
 
     def to_botorch(
-        self, parameters: list[NumericalContinuousParameter], idx_offset: int = 0
+        self, parameters: Sequence[NumericalContinuousParameter], idx_offset: int = 0
     ) -> tuple[Tensor, Tensor, float]:
         """Cast the constraint in a format required by botorch.
 
@@ -163,6 +163,8 @@ class ContinuousConstraint(Constraint, ABC):
             The tuple required by botorch.
         """
         import torch
+
+        from baybe.utils.torch import DTypeFloatTorch
 
         param_names = [p.name for p in parameters]
         param_indices = [