+++ baybe/searchspace/continuous.py
@@ -2,8 +2,8 @@
 
 from __future__ import annotations
 
-from collections.abc import Collection
-from typing import Any, Optional
+from collections.abc import Collection, Sequence
+from typing import Any, Optional, cast
 
 import numpy as np
 import pandas as pd
@@ -14,9 +14,11 @@ from baybe.constraints import (
     ContinuousLinearInequalityConstraint,
 )
 from baybe.parameters import NumericalContinuousParameter
+from baybe.parameters.base import ContinuousParameter
 from baybe.parameters.utils import get_parameters_from_dataframe
 from baybe.searchspace.validation import validate_parameter_names
 from baybe.serialization import SerialMixin, converter, select_constructor_hook
+from baybe.utils.basic import to_tuple
 from baybe.utils.dataframe import pretty_print_df
 from baybe.utils.numerical import DTypeFloatNumpy
 
@@ -30,16 +32,18 @@ class SubspaceContinuous(SerialMixin):
     parameter views.
     """
 
-    parameters: list[NumericalContinuousParameter] = field(
-        validator=lambda _1, _2, x: validate_parameter_names(x)
+    parameters: tuple[NumericalContinuousParameter, ...] = field(
+        converter=to_tuple, validator=lambda _, __, x: validate_parameter_names(x)
     )
     """The list of parameters of the subspace."""
 
-    constraints_lin_eq: list[ContinuousLinearEqualityConstraint] = field(factory=list)
+    constraints_lin_eq: tuple[ContinuousLinearEqualityConstraint, ...] = field(
+        converter=to_tuple, factory=tuple
+    )
     """List of linear equality constraints."""
 
-    constraints_lin_ineq: list[ContinuousLinearInequalityConstraint] = field(
-        factory=list
+    constraints_lin_ineq: tuple[ContinuousLinearInequalityConstraint, ...] = field(
+        converter=to_tuple, factory=tuple
     )
     """List of linear inequality constraints."""
 
@@ -92,7 +96,7 @@ class SubspaceContinuous(SerialMixin):
 
         # Create the corresponding parameters and from them the search space
         parameters = [
-            NumericalContinuousParameter(name, bound)
+            NumericalContinuousParameter(cast(str, name), bound)
             for (name, bound) in bounds.items()
         ]
         return SubspaceContinuous(parameters)
@@ -101,7 +105,7 @@ class SubspaceContinuous(SerialMixin):
     def from_dataframe(
         cls,
         df: pd.DataFrame,
-        parameters: Optional[list[NumericalContinuousParameter]] = None,
+        parameters: Optional[Sequence[ContinuousParameter]] = None,
     ) -> SubspaceContinuous:
         """Create a hyperrectangle-shaped continuous subspace from a dataframe.
 
@@ -119,10 +123,23 @@ class SubspaceContinuous(SerialMixin):
                 is created with default optional arguments. For more details, see
                 :func:`baybe.parameters.utils.get_parameters_from_dataframe`.
 
+        Raises:
+            ValueError: If parameter types other than
+                :class:`baybe.parameters.numerical.NumericalContinuousParameter`
+                are provided.
+
         Returns:
             The created continuous subspace.
         """
         # TODO: Add option for convex hull once constraints are in place
+
+        if parameters and not all(
+            isinstance(p, NumericalContinuousParameter) for p in parameters
+        ):
+            raise ValueError(
+                "Currently, only parameters of type "
+                "'{NumericalContinuousParameter.__name__}' are supported."
+            )
 
         def continuous_parameter_factory(name: str, values: Collection[Any]):
             return NumericalContinuousParameter(name, (min(values), max(values)))
@@ -140,9 +157,9 @@ class SubspaceContinuous(SerialMixin):
         return len(self.parameters) == 0
 
     @property
-    def param_names(self) -> list[str]:
+    def param_names(self) -> tuple[str, ...]:
         """Return list of parameter names."""
-        return [p.name for p in self.parameters]
+        return tuple(p.name for p in self.parameters)
 
     @property
     def param_bounds_comp(self) -> np.ndarray: