+++ CHANGELOG.md
@@ -4,9 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
 ### Added
+- `mypy` for search space
 - Class hierarchy for objectives
+
+### Changed
+- `torch` numeric types are now loaded lazily
+- Reorganized acquisition.py into `acquisition` subpackage
+- `torch` is imported lazily in `surrogates`
+
+### Fixed
+- `n_task_params` now evaluates to 1 if `task_idx == 0`
 
 ### Deprecations
 - The former `baybe.objective.Objective` class has been replaced with
@@ -41,7 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Full lookup backtesting example now tests different substance encodings
 - Replaced unmaintained `mordred` dependency by `mordredcommunity`
-- `SearchSpace`s now use `ndarray` instead of `Tensor` 
+- `SearchSpace`s now use `ndarray` instead of `Tensor`
 
 ### Fixed
 - `from_simplex` now efficiently validated in `Campaign.validate_config`